秒杀

+++ star-management/management-business/src/main/java/com/starnft/star/business/domain/WhiteListDetail.java
@@ -25,7 +25,7 @@ public class WhiteListDetail  implements Serializable {
      * 白名单id
      */
     @ApiModelProperty(name = "白名单id", notes = "")
-    private Integer whiteId;
+    private Long whiteId;
     /**
      * 用户id
      */
